move allocations to tease folder
