+++ servers/team_allocation/tease/router.go
@@ -8,6 +8,7 @@ import (
 	"github.com/google/uuid"
 	promptSDK "github.com/ls1intum/prompt-sdk"
 	"github.com/ls1intum/prompt-sdk/keycloakTokenVerifier"
+	"github.com/ls1intum/prompt2/servers/team_allocation/tease/teaseDTO"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -20,6 +21,10 @@ func setupTeaseRouter(routerGroup *gin.RouterGroup, authMiddleware func(allowedR
 	// course phase specific endpoints
 	teaseCoursePhaseRouter := teaseRouter.Group("/course_phase/:coursePhaseID")
 	teaseCoursePhaseRouter.GET("/students", authMiddleware(promptSDK.PromptAdmin, promptSDK.CourseLecturer), getTeaseStudentsForCoursePhase)
+
+	teaseCoursePhaseRouter.PUT("/allocations", authMiddleware(promptSDK.PromptAdmin, promptSDK.CourseLecturer), putAllocation)
+	teaseCoursePhaseRouter.GET("/allocations/:courseParticipationID", authMiddleware(promptSDK.PromptAdmin, promptSDK.CourseLecturer), getStudentAllocation)
+	teaseCoursePhaseRouter.GET("/allocations", authMiddleware(promptSDK.PromptAdmin, promptSDK.CourseLecturer), getAllAllocations)
 }
 
 func getAllCoursePhases(c *gin.Context) {
@@ -69,4 +74,59 @@ func getTeaseStudentsForCoursePhase(c *gin.Context) {
 
 func handleError(c *gin.Context, statusCode int, err error) {
 	c.JSON(statusCode, gin.H{"error": err.Error()})
+}
+
+func getAllAllocations(c *gin.Context) {
+	coursePhaseID, err := uuid.Parse(c.Param("coursePhaseID"))
+	if err != nil {
+		handleError(c, http.StatusBadRequest, errors.New("invalid course phase ID"))
+		return
+	}
+
+	allocations, err := GetAllocationsByCoursePhase(c, coursePhaseID)
+	if err != nil {
+		handleError(c, http.StatusInternalServerError, err)
+		return
+	}
+
+	c.JSON(http.StatusOK, allocations)
+}
+
+func getStudentAllocation(c *gin.Context) {
+	courseParticipationID, err := uuid.Parse(c.Param("courseParticipationID"))
+	if err != nil {
+		handleError(c, http.StatusBadRequest, err)
+		return
+	}
+
+	coursePhaseID, err := uuid.Parse(c.Param("coursePhaseID"))
+	if err != nil {
+		handleError(c, http.StatusBadRequest, errors.New("invalid course phase ID"))
+		return
+	}
+
+	allocation, err := GetStudentAllocation(c, courseParticipationID, coursePhaseID)
+	if err != nil {
+		handleError(c, http.StatusInternalServerError, err)
+		return
+	}
+
+	c.JSON(http.StatusOK, allocation)
+}
+
+func putAllocation(c *gin.Context) {
+	var req teaseDTO.AllocationRequest
+
+	if err := c.BindJSON(&req); err != nil {
+		handleError(c, http.StatusBadRequest, err)
+		return
+	}
+
+	err := PutAllocation(c, req.CourseParticipationID, req.TeamID, req.CoursePhaseID)
+	if err != nil {
+		handleError(c, http.StatusInternalServerError, err)
+		return
+	}
+
+	c.JSON(http.StatusCreated, gin.H{"message": "Allocation created successfully"})
 }
