+++ servers/team_allocation/tease/teaseDTO/allocationRequest.go
@@ -1,4 +1,4 @@
-package allocationDTO
+package teaseDTO
 
 import (
 	"github.com/google/uuid"
